feature: Added functionality for CEO employment term increase/decrease. Fixed a bug with CEO instance on Company assignment. Updated CompanyLevel with chance to affect CEO. Updated CEO employment desire so that Great CEOs will leave sooner and Terrible ones will stay longer.

+++ Assets/_Project/Scripts/Constants/CEOLevel.cs
@@ -2,13 +2,14 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-[CreateAssetMenu(fileName = "CEOLevel", menuName = "AdventuresOnWallStreet/CEOLevel", order = 1)]
-public class CEOLevel : ScriptableObject
-{
+[CreateAssetMenu (fileName = "CEOLevel", menuName = "AdventuresOnWallStreet/CEOLevel", order = 1)]
+public class CEOLevel : ScriptableObject {
+    // Employment days are what the CEO prefers and not what the company is likely to want, so a terrible CEO will try to stay longer since options are fewer.
+    // TODO: need to create the % chance to go up or down a level on re-hire.
     // Ranges
     public float compentcyScoreMinimum;
     public float compentcyScoreMaximum;
-    public float employmentDaysMinimum;
-    public float employmentDaysMaximum;
+    public int employmentDaysMinimum;
+    public int employmentDaysMaximum;
     public GameLevels gameLevel;
 }
